refactor(storage): extract usage stats store
